ログインユーザーと出品者を判断する

+++ app/controllers/items_controller.rb
@@ -38,9 +38,11 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+    if current_user.id == @item.user_id
+      item = Item.find(params[:id])
+      item.destroy
+      redirect_to root_path
+    end
   end
 
   private
